refactor: 아키텍처 규칙 위반 수정
- ImageUploadStatus를 domain -> constant 패키지로 이동
- ImageUploadLogItem을 dto -> dto.response 패키지로 이동
- 관련 import 문 수정
