+++ bottlenote-mono/src/main/java/app/bottlenote/common/file/service/ImageUploadLogService.java
@@ -1,10 +1,10 @@
 package app.bottlenote.common.file.service;
 
+import app.bottlenote.common.file.constant.ImageUploadStatus;
 import app.bottlenote.common.file.domain.ImageUploadLog;
 import app.bottlenote.common.file.domain.ImageUploadLogRepository;
-import app.bottlenote.common.file.domain.ImageUploadStatus;
-import app.bottlenote.common.file.dto.ImageUploadLogItem;
 import app.bottlenote.common.file.dto.request.ImageUploadLogRequest;
+import app.bottlenote.common.file.dto.response.ImageUploadLogItem;
 import app.bottlenote.common.file.dto.response.ImageUploadLogResponse;
 import java.time.LocalDateTime;
 import java.util.List;
